create endpoint medical record list
Created the endpoint to list all patients, including name, patient_id and insurance, with pagination. Classes related to the Patient were used to handle this endpoint, as it returns data related to the patient. However, 2 new DTOs were created for this endpoint. The logic for using filter parameters was included in the repository.