create endpoint medical record list, add new responses to existing endpoints
Adds the new medical record list endpoint, which queries all patients and returns each patient's ID, name and insurance. It is possible to add parameters to this endpoint and filter only the desired patient, using id or name. If no parameters are provided, it will return all patients with paging. All endpoints already created in the project were retested and new customized responses to queries were added. A series of details that were identified during tests for the correct functioning of the endpoints were also corrected.