create patient medical record endpoint
Add patient medical record endpoint, where it is possible to consult the medical record of a specific patient whose ID is provided. In the medical record, some patient data is returned, as well as all exams and medical appointments linked to him. All endpoints already created in the project were retested and new custom responses to queries were added. A series of details that were identified during testing for the correct functioning of the endpoints were also corrected.