fix permission logic for patients
Fixed method that extracts user id from jwt token. Moved all the logic that involves the patient's permission to obtain only their own medical record based on their id as a patient to the medical record controller, as within SecurityConfig it was not possible to retrieve the path variables for security reasons within spring security itself, which protected information before it could be obtained.